# backend/h2os/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/employers/{id}:
    get:
      tags:
      - employer-rest-controller
      summary: Get an employer by its id
      operationId: getEmployer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found the employer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployerDTO'
        "404":
          description: "Employer not found, probably invalid id supplied"
    put:
      tags:
      - employer-rest-controller
      summary: Update an employer fields by ID. You need to be an administrator or
        the own employer.
      operationId: updateEmployer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployerDTO'
        required: true
      responses:
        "400":
          description: Data entered incorrectly.
        "200":
          description: Employer updated correctly. You can update only the fields
            you want
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployerDTO'
        "404":
          description: "Employer not found, probably invalid id supplied"
        "401":
          description: You are not authorized
    delete:
      tags:
      - employer-rest-controller
      summary: Delete an employer by its id. You need to be an administrator or the
        own employer.
      operationId: deleteEmployer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Employer deleted
        "404":
          description: "Employer not found, probably invalid id supplied"
        "401":
          description: You are not authorized
...
  /api/offers:
    get:
      tags:
      - offer-rest-controller
      summary: Get paged offers.
      operationId: getOffers
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 5
      responses:
        "204":
          description: "Offers not found, probably high page number supplied"
        "200":
          description: Offers found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferDTO'
    post:
      tags:
      - offer-rest-controller
      summary: Create an offer.
      operationId: createOffer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfferDTO'
        required: true
      responses:
        "404":
          description: "Pool not found, probably invalid id supplied"
        "201":
          description: Offer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferDTO'
        "401":
          description: "You are not authorized, you are not a an employer"
        "400":
          description: "Bad request, invalid data in new offer"
...
  /api/employers:
    get:
      tags:
      - employer-rest-controller
      summary: Get paged employers.
      operationId: getEmployers
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 5
      responses:
        "200":
          description: Employers found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployerDTO'
        "204":
          description: "Employers not found, probably high page number supplied"
    post:
      tags:
      - employer-rest-controller
      summary: Post a new employer
      operationId: createEmployer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployerDTO'
        required: true
      responses:
        "400":
          description: Data entered incorrectly.
        "200":
          description: Employer posted correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployerDTO'
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Employer'
  /api/employers/{id}/photoCompany:
    get:
      tags:
      - employer-rest-controller
      summary: Get the photo of the company of an employer by id
      operationId: downloadPhotoCompany
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found the company photo
        "400":
          description: "Data entered incorrectly, probably invalid id supplied"
        "404":
          description: That employer does not have a company photo
    post:
      tags:
      - employer-rest-controller
      summary: Post a new photo of the company of an employer by id
      operationId: uploadPhotoCompany
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "201":
          description: Company photo posted correctly
        "404":
          description: "Data entered incorrectly, probably invalid id supplied"
        "401":
          description: You are not authorized
    delete:
      tags:
      - employer-rest-controller
      summary: Delete the photo of the company of an employer by id. You need to be
        an admin or the own employer
      operationId: deletePhotoCompany
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: "Data entered incorrectly, probably invalid id supplied"
        "200":
          description: Company photo deleted correctly
        "401":
          description: You are not authorized
...
  /api/employers/{id}/offers:
    get:
      tags:
      - employer-rest-controller
      summary: Get the offers of an employer by id. You need to be an admin or the
        own employer
      operationId: getEmployerOffers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found the offer list of the employer
        "404":
          description: "Data entered incorrectly, probably invalid id supplied"
        "401":
          description: You are not authorized
...
  /api/employers/{id}/offers/{offerId}:
    delete:
      tags:
      - employer-rest-controller
      summary: Delete an offer of the employer by id. You need to be an admin or the
        own employer
      operationId: deleteOfferFromEmployer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: offerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: "Data entered incorrectly, probably invalid id or offerId supplied"
        "200":
          description: Offer deleted correctly
        "401":
          description: You are not authorized
components:
  schemas:
...
    OfferDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        poolID:
          type: integer
          format: int64
        poolName:
          type: string
        direction:
          type: string
        salary:
          type: string
        type:
          type: string
        start:
          type: string
        description:
          type: string
        employer:
          type: string
        employerID:
          type: integer
          format: int64
...
    EmployerDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        description:
          type: string
        dni:
          type: string
        mail:
          type: string
        age:
          type: string
        pass:
          type: string
        phone:
          type: string
        country:
          type: string
        locality:
          type: string
        province:
          type: string
        direction:
          type: string
        type:
          type: string
        company:
          type: string
        imageCompany:
          type: boolean
    Employer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        description:
          type: string
        dni:
          type: string
        mail:
          type: string
        age:
          type: string
        phone:
          type: string
        country:
          type: string
        locality:
          type: string
        province:
          type: string
        direction:
          type: string
        roles:
          type: array
          items:
            type: string
        imageCompany:
          type: boolean
        company:
          type: string
        offers:
          type: array
          items:
            $ref: '#/components/schemas/Offer'
        photoCompany:
          type: object
          properties:
            binaryStream:
              type: object
        type:
          type: string
        admin:
          type: boolean
        address:
          type: string
    Lifeguard:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        description:
          type: string
        dni:
          type: string
        mail:
          type: string
        age:
          type: string
        phone:
          type: string
        country:
          type: string
        locality:
          type: string
        province:
          type: string
        direction:
          type: string
        roles:
          type: array
          items:
            type: string
        imageUser:
          type: boolean
        document:
          type: string
        skills:
          type: array
          items:
            type: string
        pools:
          type: array
          items:
            $ref: '#/components/schemas/Pool'
        offers:
          type: array
          items:
            $ref: '#/components/schemas/Offer'
        getofferAssigned:
          type: boolean
        photoUser:
          type: object
          properties:
            binaryStream:
              type: object
        offersAccepted:
          type: array
          items:
            $ref: '#/components/schemas/Offer'
        type:
          type: string
        admin:
          type: boolean
        address:
          type: string
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
    Message:
      type: object
      properties:
        id:
          type: integer
          format: int64
        pool:
          $ref: '#/components/schemas/Pool'
        author:
          type: string
        body:
          type: string
        owner:
          type: boolean
    Offer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        pool:
          $ref: '#/components/schemas/Pool'
        employer:
          $ref: '#/components/schemas/Employer'
        lifeguard:
          $ref: '#/components/schemas/Lifeguard'
        lifeguards:
          type: array
          items:
            $ref: '#/components/schemas/Lifeguard'
        salary:
          type: string
        start:
          type: string
        type:
          type: string
        description:
          type: string
        acceptedByProfileUser:
          type: boolean
        accepted:
          type: boolean
    Pool:
      type: object
      properties:
        id:
          type: integer
          format: int64
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        name:
          type: string
        photo:
          type: string
        photoBlob:
          type: object
          properties:
            binaryStream:
              type: object
        direction:
          type: string
        capacity:
          type: integer
          format: int32
        scheduleStart:
          $ref: '#/components/schemas/LocalTime'
        scheduleEnd:
          $ref: '#/components/schemas/LocalTime'
        company:
          type: string
        description:
          type: string
        end:
          $ref: '#/components/schemas/LocalTime'
        start:
          $ref: '#/components/schemas/LocalTime'
...